+++ client/Packages/com.beamable/Editor/BeamEditor.cs
@@ -41,6 +41,8 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using Beamable.Server.Editor;
+using Beamable.Server.Editor.Usam;
 using UnityEditor;
 using UnityEditor.AddressableAssets;
 using UnityEngine;
@@ -125,6 +127,10 @@ namespace Beamable
 			DependencyBuilder.AddSingleton<DotnetService>();
 			DependencyBuilder.AddSingleton<IRuntimeConfigProvider>(p => new EditorRuntimeConfigProvider(p.GetService<AccountService>()));
 
+			DependencyBuilder.AddGlobalStorage<UsamService, SessionStorageLayer>();
+			DependencyBuilder.AddSingleton(() => MicroserviceConfiguration.Instance);
+			DependencyBuilder.AddSingleton<CommonAreaService>();
+			
 			DependencyBuilder.AddSingleton<SingletonDependencyList<ILoadWithContext>>();
 			OpenApiRegistration.RegisterOpenApis(DependencyBuilder);
 		}